requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-dataproc==4.0.2
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.0
py4j==0.10.9.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
